[Clean & Add] : Add comments on what should be found in each file and cleaning of the repo
