+++ jigsolver/__init__.py
@@ -1,5 +1,5 @@
 from .puzzle import Board, Border, Piece, Puzzle, Slot
-from .random_solver import random_solver
-from .placer import naiv_greedy_placer
-from .metrics import cho_CM,pomeranz_CM,BestBuddies_matrix
-from .segmenter import segmenter,find_segment,remove_all_but_segment
+from jigsolver.random_solver import *
+from jigsolver.pomeranz_solver import *
+from .metrics import cho_CM,pomeranz_CM
+
