+++ jigsolver/random_solver/__init__.py
@@ -1,0 +1,1 @@
+import numpy as np
